requirements.txt
# Core
python-dotenv>=1.0
duckdb>=0.10.0
pandas>=2.0
numpy>=1.24
PyYAML>=6.0

# Dashboard
dash>=2.14
dash-bootstrap-components>=1.5
plotly>=5.18

# Reports
reportlab>=4.0
Jinja2>=3.1
kaleido>=0.2

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0
